trace2hooks: Rename pack-objects metric
The pack-objects metric that tracks elapsed time per pack-objects stage
is helpfully called "metric". This is okay-ish in a world where we only
have one metric we want to track, but in a subsequent commit we're about
to introduce a second metric.

Rename it to "stagesHistogram" so that it can more readily be told apart
from the upcoming second metric.