trace2hooks: Track metric label and log field in the same struct
We have two structures `statsElapsedTimes` and `histogramStageNames`,
both of which are keyed by the same Git trace2 event names. This results
in a somewhat repetitive pattern where we have to verify twice that the
given event exists, even though we know that we want to log each of
these events both via Prometheus and our custom log fields.

Unify the metric label and log field in a structure to simplify this a
bit.