trace2hooks: Record pack-objects stats via Prometheus metrics
Next to the time spent in specific stages of git-pack-objects(1) we also
track some statistics around the underpinnings of it, like for example
how many objects we've written as part of the pack. But in contrast to
the stage timings, this data is only accessible via log fields and not
via Prometheus. This makes it hard to monitor those stats on a global
scale.

Start to collect those statistics via Prometheus to plug this gap.