trace2hooks: Use "written" instead of "write_pack_file/wrote"
The trace2 data of git-pack-objects(1) yields the number of objects via
two equivalent fields: "written" and "write_pack_file/wrote". While
we're currently using the latter data field, we're about to extend the
statistics to also cover more metrics around how many objects were
written, and all of them fit in with the former of these two data
fields. This will result in some small inconsistencies and puzzlement.

Adapt the code to use "written" instead to prepare for this.

Note that we have to adapt some of our tests to reflect this change.
This may seem brittle, but we also have some additional tests in
"packobjects_metrics_test_ext.go" that exercise this logic with an
actual git-pack-objects(1) process. Those tests don't have to be
adapted, which demonstrates that both of these fields really are
equivalent.