featureflag: Default-enable multi-pack reuse
When generating packfiles, Git tries to reuse objects and their deltas
to the best extent possible so that it can avoid recomputing them. In
the best case, the repository has bitmap coverage of a significant part
of the packfile and. If that is the case, then Git can compute
reachability of objects via those bitmaps and then directly stream all
objects marked as reachable via the bitmap as-is from the packfile. This
is called "pack-reuse".

For a long time, Git only knew how to do this for a single packfile. But
with the introduction of multi-pack indices and their bitmap extension
we now have the ability to have bitmaps that cover multiple packs. With
this infra, it is thus theoretically possible to stream objects as-is
from multiple packfiles.

This feature has been introduced into Git with af626ac0e0 (pack-bitmap:
enable reuse from all bitmapped packs, 2023-12-14), but it exists as an
opt-in feature. In Gitaly 18.1, we introduced a feature flag to make us
set the required configuration.

But for a long time I have been puzzled why this setting isn't working
at all. As it turns out, it actually _is_ working, but there were some
circumstances that made it seem like it's not:

  - I have mostly been testing with gitlab-org/gitlab. But that
    repository is a pooled repository, and the pool has not been updated
    since March 2025. Because of that, we only got a single packfile in
    there with bitmap coverage, so naturally multi-pack reuse cannot do
    anything.

  - The heuristic for how we regenerate the multi-pack index is quite
    strict, so we only update the MIDX rather infrequently. As
    multi-pack reuse depends on the MIDX this means that we often have a
    MIDX that only covers a single packfile.

Together, my shallow testing seemed to indicate things as not working,
even though that's not the case.

We have since implemented better monitoring for this scenario to give us
better visibility going forward. So let's enable the feature by removing
the flag -- we have already been testing the feature for ~9 months in
production, and it never caused any issues.