transactions: Add config param for dry-run mode
We previously enabled dry-run middleware by default and removed the
usage of feature flag. This caused confusion for some customers due
to sudden appearance of dry-run logs as well as small performance
related overheads for very specific workflows.
Dry-run middeware has a cache mechanism to prevent any performance
impact during regular traffic, but for certain workflows such as
collecting statistics for entire projects on a node can now have
increase in duration due to this middleware as reported here:
https://gitlab.com/gitlab-org/gitaly/-/issues/7018

We should make this middleware controlleable again. Therefore we
are reintroducing the feature flag for this feature.